requirements.txt
Django==5.1.6
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
flake8==7.0.0
black==24.1.1
isort==5.13.2
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
django-debug-toolbar==4.2.0
